requirements.txt
asgiref==3.3.4
cffi==1.15.1
click==8.0.1
colorama==0.4.4
cryptography==37.0.4
ecdsa==0.17.0
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
pip==20.0.2
pkg-resources==0.0.0
psycopg2-binary==2.9.1
pyasn==10.4.8
pycparser==2.21
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
setuptools==44.0.0
six==1.16.0
SQLAlchemy==1.4.19
SQLAlchemy-Utils==0.38.2
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
bcrypt==3.2.2
passlib==1.7.4